Move Kokkos_SetEnv.hpp to unit_test directory
